+++ CHANGELOG.md
@@ -26,6 +26,11 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 * `jj show` now supports `--reversed` flag.
 
+* `jj run` allows you to run a command over a set of changes, each with their
+  own private working copy; the commands may update the working copy and
+  changes/conflicts are propagated accordingly, e.g., `jj run -- cargo check
+  --all-features` or `jj run -- cargo fix` behaves as one might expect.
+
 ### Fixed bugs
 
 * `jj` now creates a new working-copy revision during snapshotting if the